+++ src/jtl/Connector/Shopware/Controller/CustomerOrder.php
@@ -9,6 +9,7 @@ namespace jtl\Connector\Shopware\Controller;
 use jtl\Connector\Core\Utilities\Money;
 use jtl\Connector\Formatter\ExceptionFormatter;
 use jtl\Connector\Model\Identity;
+use jtl\Connector\Payment\PaymentTypes;
 use jtl\Connector\Result\Action;
 use jtl\Connector\Shopware\Utilities\Locale as LocaleUtil;
 use jtl\Connector\Shopware\Utilities\Mmc;
@@ -58,6 +59,18 @@ class CustomerOrder extends DataController
                     $paymentModuleCode = PaymentUtil::map(null, $orderSW['payment']['name']);
                     $paymentModuleCode = ($paymentModuleCode !== null) ? $paymentModuleCode : $orderSW['payment']['name'];
                     $order->setPaymentModuleCode($paymentModuleCode);
+
+                    // Billsafe
+                    if ($paymentModuleCode === PaymentTypes::TYPE_BILLSAFE
+                    && isset($orderSW['attribute']['swagBillsafeIban'])
+                    && isset($orderSW['attribute']['swagBillsafeBic'])) {
+                        $order->setPui(sprintf(
+                            'Bitte bezahlen Sie %s %s an folgendes Konto: %s',
+                            $orderSW['invoiceAmount'],
+                            $order->getCurrencyIso(),
+                            sprintf('IBAN: %s, BIC: %s', $orderSW['attribute']['swagBillsafeIban'], $orderSW['attribute']['swagBillsafeBic'])
+                        ));
+                    }
 
                     // CustomerOrderStatus
                     $customerOrderStatus = StatusUtil::map(null, $orderSW['status']);
@@ -141,18 +154,17 @@ class CustomerOrder extends DataController
                     }
 
                     // Adding shipping item
-                    if ($orderSW['invoiceShippingNet'] > 0) {
-                        $item = Mmc::getModel('CustomerOrderItem');
-                        $item->setType(\jtl\Connector\Model\CustomerOrderItem::TYPE_SHIPPING)
-                            ->setId(new Identity(sprintf('%s_ship', $orderSW['id'])))
-                            ->setCustomerOrderId($order->getId())
-                            ->setName('Shipping')
-                            ->setPrice($orderSW['invoiceShippingNet'])
-                            ->setQuantity(1)
-                            ->setVat(self::calcShippingVat($order));
+                    $shippingPrice = (isset($orderSW['invoiceShippingNet'])) ? (float) $orderSW['invoiceShippingNet'] : 0.0;
+                    $item = Mmc::getModel('CustomerOrderItem');
+                    $item->setType(\jtl\Connector\Model\CustomerOrderItem::TYPE_SHIPPING)
+                        ->setId(new Identity(sprintf('%s_ship', $orderSW['id'])))
+                        ->setCustomerOrderId($order->getId())
+                        ->setName('Shipping')
+                        ->setPrice($shippingPrice)
+                        ->setQuantity(1)
+                        ->setVat(self::calcShippingVat($order));
 
-                        $order->addItem($item);
-                    }
+                    $order->addItem($item);
 
                     // Attributes
                     for ($i = 1; $i <= 6; $i++) {