+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.3.0
+-----
+- Added connector install, phar and suhosin check
+- Added category language locale check
+- Added billsafe to payment types
+- Added billsafe customer order support
+- Added shopware 5.1 media service support
+- Added file and mysql validation files
+- Changed payment mysql trigger to trigger only at state 'payment completed'
+- Changed customer order shipping item support
+
 1.2.5
 -----
 - Fixed product attribute type